Update testit.bat
Blows away Sonarlint config directory if testing being done for the first time on this project - eliminates some issues with Sonarlint crapping out if old cache files are present. 

Also, speed up Sonarlint a bit by removing the -u flag from the call; by doing so, if this is the first check being run, the server will be contacted and rules will be cached locally, but if it's not the first run, the cached rules will be used.